refactor(collapse): script setup
