+++ src/packages/__VUE/collapse/types.ts
@@ -1,6 +1,8 @@
 import type { InjectionKey } from 'vue';
 
 export const COLLAPSE_KEY = Symbol('nut-collapse') as InjectionKey<{
-  updateVal: Function;
-  isExpanded: Function;
+  updateVal: (name: string | number) => void;
+  isExpanded: (name: string | number) => boolean;
 }>;
+
+export type CollapseValue = string | number | (string | number)[];
